Module mysql::prelude

Traits

AsStatement

Something, that eventually is a Statement in the context of a T: Queryable.

BatchQuery

Helper trait for batch statement execution.

BinQuery

MySql prepared statement query.

ColumnIndex

Things that may be used as an index of a row column.

ConvIr

Basic operations on FromValue conversion intermediate result.

FromRow

Trait to convert Row into a tuple of FromValue implementors up to arity 12.

FromValue

Implement this trait to convert value to something.

Protocol

Trait for protocol markers crate::Binary and crate::Text.

Queryable

Queryable object.

TextQuery

MySql text query.

ToValue

WithParams

Helper, that constructs QueryWithParams.