Struct mysql::PooledConn

pub struct PooledConn { /* fields omitted */ }

Pooled mysql connection which will return to the pool on drop.

You should prefer using prepare or prep_exec instead of query where possible, except cases when statement has no params and when it has no return values or return values which evaluates to Value::Bytes.

query is a part of mysql text protocol, so under the hood you will always receive Value::Bytes as a result and from_value will need to parse it if you want, for example, i64

let pool = Pool::new(get_opts()).unwrap();
let mut conn = pool.get_conn().unwrap();

conn.query_first("SELECT 42").map(|result: Option<Value>| {
    let result = result.unwrap();
    assert_eq!(result, Value::Bytes(b"42".to_vec()));
    assert_eq!(from_value::<i64>(result), 42i64);
}).unwrap();
conn.exec_iter("SELECT 42", ()).map(|mut result| {
    let cell = result.next().unwrap().unwrap().take(0).unwrap();
    assert_eq!(cell, Value::Int(42i64));
    assert_eq!(from_value::<i64>(cell), 42i64);
}).unwrap();

For more info on how to work with query results please look at QueryResult documentation.

Implementations

impl PooledConn

pub fn start_transaction(&mut self, tx_opts: TxOpts) -> Result<Transaction<'_>>

Redirects to Conn#start_transaction

pub fn as_mut(&mut self) -> &mut Conn

Gives mutable reference to the wrapped Conn.

pub fn as_ref(&self) -> &Conn

Gives reference to the wrapped Conn.

pub fn unwrap(self) -> Conn

Unwraps wrapped Conn.

pub fn set_local_infile_handler(&mut self, handler: Option<LocalInfileHandler>)

A way to override default local infile handler for this pooled connection. Destructor will restore original handler before returning connection to a pool. See Conn::set_local_infile_handler.