Module mysql::value

Structs

BytesIr

OptionIr

ParseIr

StringIr

ValueIr

Enums

Params

Value
Value enumerates possible values in mysql cells. Also Value used to fill prepared statements.

Traits

ConvIr
Basic operations on FromValue conversion intermediate result.

FromRow
Trait to convert Row into tuple of FromValue implementors up to arity 12.

FromValue
Implement this trait to convert value to something.

ToValue

Functions

from_row
Will panic if could not convert row to T.

from_row_opt
Will return Err(row) if could not convert row to T

from_value
Will panic if could not convert v to T

from_value_opt
Will return Err(v) if could not convert v to T