Crate mysql

rust-mysql-simple

Mysql client library implemented in rust nightly.

Install

Just include another [dependencies.*] section into your Cargo.toml:

[dependencies.mysql]
git = "https://github.com/blackbeam/rust-mysql-simple"

rust-mysql-simple offer support of SSL via ssl cargo feature which is enabled by default. If you have no plans to use SSL, then you should disable that feature to not to depend on rust-openssl:

[dependencies.mysql]
git = "https://github.com/blackbeam/rust-mysql-simple"
default-features = false

Use

You should start by creating MyOpts struct.

Then you can create MyPool which should be enough to work with mysql server.

Modules

conn

consts

error

value