Struct mysql::conn::Stmt

pub struct Stmt<'a> {
    // some fields omitted
}

Mysql prepared statement.

Methods

impl<'a> Stmt<'a>

fn params_ref(&self) -> Option<&[Column]>

Returns a slice of a Columns which represents Stmt's params if any.

fn columns_ref(&self) -> Option<&[Column]>

Returns a slice of a Columns which represents Stmt's columns if any.

fn column_index<T: AsRef<str>>(&self, name: T) -> Option<usize>

Returns index of a Stmt's column by name.

fn execute<'s>(&'s mut self, params: &[&ToValue]) -> MyResult<QueryResult<'s>>

Executes prepared statement with an arguments passed as a slice of a references to a ToValue trait implementors.