Module mysql::value

Enums

Value
Value enumerates possible values in mysql cells. Also Value used to fill prepared statements. Note that to receive something other than Value::NULL or Value::Bytes from mysql you should use prepared statements.

Traits

FromValue

ToValue

Functions

from_value
Will panic if could not retrieve Self from Value

from_value_opt
Will return None if could not retrieve Self from Value